Revert "fix(product-analytics): fit journey ribbon stacks within their cards"
This reverts commit 4283b9648f1a3e7911e2939b76c33bc0dbe83d94, keeping
the grid rendering as it is on the combined branch. The anchor editor
and drop-off tooltip changes stay.

Generated-By: PostHog Code
Task-Id: fc078ac7-28e3-4c04-99c5-4f5d1cab52bf
