+++ products/product_analytics/frontend/insights/journeys/JourneyGrid.tsx
@@ -9,7 +9,6 @@ import { midEllipsis, pluralize } from 'lib/utils/strings'
 
 import { PathsV2AnchorType } from '~/queries/schema/schema-general'
 
-import { CARD_HEIGHT, CARD_WIDTH, COLUMN_GAP, cardKey, computeJourneyGridGeometry } from './journeyGridGeometry'
 import {
     JourneyChainHighlight,
     JourneyGridModel,
@@ -19,10 +18,39 @@ import {
     ribbonPath,
 } from './journeyGridModel'
 
+const CARD_WIDTH = 224
+const COLUMN_GAP = 104
+const CARD_HEIGHT = 66
+const ROW_GAP = 14
+const HEADER_HEIGHT = 28
+const PORT_TOP_OFFSET = 6
+const PORT_GAP = 2
+const MAX_RIBBON_THICKNESS = 36
+const MIN_RIBBON_THICKNESS = 2
 const MAX_LABEL_CHARS = 30
 const RIBBON_OPACITY = 0.15
 const RIBBON_OPACITY_ON_CHAIN = 0.4
 const RIBBON_OPACITY_DIMMED = 0.06
+
+interface CardGeometry {
+    row: JourneyGridRow
+    stepIndex: number
+    x: number
+    y: number
+}
+
+interface RibbonGeometry {
+    ribbon: JourneyGridRibbon
+    thickness: number
+    fromX: number
+    fromY: number
+    toX: number
+    toY: number
+}
+
+function cardKey(stepIndex: number, rowKey: string): string {
+    return `${stepIndex}:${rowKey}`
+}
 
 /** End-anchored grids read backward in time from the anchor column, so "Step n" would label the
  * columns in the wrong direction. */
@@ -56,7 +84,59 @@ export function JourneyGrid({
     onRibbonHover?: (ribbon: JourneyGridRibbon) => void
     onGridLeave?: () => void
 }): JSX.Element {
-    const { cards, ribbons, chartWidth, chartHeight } = useMemo(() => computeJourneyGridGeometry(model), [model])
+    const { cards, ribbons, chartWidth, chartHeight } = useMemo(() => {
+        const cardByKey = new Map<string, CardGeometry>()
+        const cards: CardGeometry[] = []
+        let maxCardBottom = 0
+        model.columns.forEach((column, columnIndex) => {
+            let y = HEADER_HEIGHT
+            for (const row of column.rows) {
+                const card = { row, stepIndex: column.stepIndex, x: columnIndex * (CARD_WIDTH + COLUMN_GAP), y }
+                cardByKey.set(cardKey(column.stepIndex, row.key), card)
+                cards.push(card)
+                y += CARD_HEIGHT + ROW_GAP
+            }
+            maxCardBottom = Math.max(maxCardBottom, y - ROW_GAP)
+        })
+
+        const thicknessFor = (count: number): number =>
+            Math.max(MIN_RIBBON_THICKNESS, (count / Math.max(model.maxRibbonCount, 1)) * MAX_RIBBON_THICKNESS)
+
+        // Resolve each ribbon's cards once, then stack the ports on each card side in the vertical
+        // order of the connected cards, so ribbons out of one card never cross at the card edge.
+        const resolvedRibbons = model.ribbons.flatMap((ribbon) => {
+            const sourceCardKey = cardKey(ribbon.sourceStep, ribbon.sourceKey)
+            const targetCardKey = cardKey(ribbon.sourceStep + 1, ribbon.targetKey)
+            const source = cardByKey.get(sourceCardKey)
+            const target = cardByKey.get(targetCardKey)
+            return source && target ? [{ ribbon, source, target, sourceCardKey, targetCardKey }] : []
+        })
+        resolvedRibbons.sort((a, b) => a.source.y - b.source.y || a.target.y - b.target.y)
+
+        const outCursor = new Map<string, number>()
+        const inCursor = new Map<string, number>()
+        const ribbons: RibbonGeometry[] = resolvedRibbons.map(
+            ({ ribbon, source, target, sourceCardKey, targetCardKey }) => {
+                const thickness = thicknessFor(ribbon.count)
+                const fromY = source.y + PORT_TOP_OFFSET + (outCursor.get(sourceCardKey) ?? 0)
+                const toY = target.y + PORT_TOP_OFFSET + (inCursor.get(targetCardKey) ?? 0)
+                outCursor.set(sourceCardKey, (outCursor.get(sourceCardKey) ?? 0) + thickness + PORT_GAP)
+                inCursor.set(targetCardKey, (inCursor.get(targetCardKey) ?? 0) + thickness + PORT_GAP)
+                return {
+                    ribbon,
+                    thickness,
+                    fromX: source.x + CARD_WIDTH,
+                    fromY,
+                    toX: target.x,
+                    toY,
+                }
+            }
+        )
+
+        const chartWidth = model.columns.length * (CARD_WIDTH + COLUMN_GAP) - COLUMN_GAP
+        const chartHeight = maxCardBottom + 16
+        return { cards, ribbons, chartWidth, chartHeight }
+    }, [model])
 
     const ribbonOpacity = (ribbon: JourneyGridRibbon): number => {
         if (!chainHighlight) {
@@ -70,7 +150,7 @@ export function JourneyGrid({
             {/* eslint-disable-next-line react/forbid-dom-props */}
             <div className="relative" style={{ width: chartWidth, height: chartHeight }}>
                 <svg width={chartWidth} height={chartHeight} className="absolute inset-0 pointer-events-none">
-                    {ribbons.map(({ ribbon, sourceThickness, targetThickness, fromX, fromY, toX, toY }) => (
+                    {ribbons.map(({ ribbon, thickness, fromX, fromY, toX, toY }) => (
                         <Tooltip
                             key={ribbon.key}
                             title={
@@ -86,7 +166,7 @@ export function JourneyGrid({
                             }
                         >
                             <path
-                                d={ribbonPath(fromX, fromY, sourceThickness, toX, toY, targetThickness)}
+                                d={ribbonPath(fromX, fromY, thickness, toX, toY, thickness)}
                                 fill={nodeColor}
                                 opacity={ribbonOpacity(ribbon)}
                                 className={`pointer-events-auto hover:opacity-40 transition-opacity ${
